rename circle_progress_bar.js to sf-circle_progress_bar.js
