+++ src/components/sf-circle_progress_bar.js
